luxebox/pyproject.toml
[project]
name = "luxebox"
version = "0.1.0"
description = "Local coding LLM evaluation + optimization harness for Apple Silicon"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.8",
    "pyyaml>=6.0",
    "rich>=13.7",
    "typer>=0.12",
    "datasets>=3.0",
    "huggingface-hub>=0.26",
    "psutil>=6.0",
    "jinja2>=3.1",
    "tenacity>=9.0",
    "tqdm>=4.66",
    "ddgs>=9.0",
    "trafilatura>=1.12",
]

[project.optional-dependencies]
mlx = [
    "mlx>=0.21",
    "mlx-lm>=0.20",
]
evalplus = [
    "evalplus>=0.3.1",
]
# `bfcl-eval` is intentionally NOT listed here.
# It pins tree-sitter==0.21.3, which conflicts with evalplus (>=0.22). Install
# BFCL into its own venv when you want to run that benchmark:
#
#   uv venv .venv-bfcl
#   .venv-bfcl/bin/pip install bfcl-eval httpx rich typer tenacity psutil pyyaml pydantic
#   .venv-bfcl/bin/python scripts/run_phase_a.py --candidate <id> --bench bfcl_v3

[project.scripts]
lux = "harness.cli:app"
luxe = "luxe.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["harness", "benchmarks", "personal_eval", "luxe"]

[tool.ruff]
line-length = 100
target-version = "py311"
